+++ src/blocks/gravity-form-renderer/Edit.tsx
@@ -1,32 +1,25 @@
 import { useBlockProps, InspectorControls } from '@wordpress/block-editor';
-import { PanelBody, PanelRow, SelectControl } from '@wordpress/components';
-import { useState, useEffect } from '@wordpress/element';
+import {
+	PanelBody,
+	PanelRow,
+	SelectControl,
+	TextControl,
+	Notice,
+	Tip,
+} from '@wordpress/components';
 
-export default function Edit( { attributes, setAttributes } ) {
+import useGravityForm from './hooks/useGravityForm';
+import ModalControls from './editorComponents/ModalControls';
+
+export default function Edit( props ) {
+	const { attributes, setAttributes } = props;
 	const blockProps = useBlockProps();
-	const [ options, setOptions ] = useState<
-		{ label: string; value: number }[]
-	>( [] );
 	const { formTitle, formId } = attributes;
+	const { isLoading, options, preFillableFields, errors } = useGravityForm(
+		formId,
+		attributes
+	);
 
-	useEffect( () => {
-		// Fetch available Gravity Forms from the REST API
-		fetch( '/wp-json/gf/v2/forms' )
-			.then( ( response ) => response.json() )
-			.then( ( data ) => {
-				const options = Object.entries( data ).map(
-					( [ , form ] ) => ( {
-						label: form.title,
-						value: +form.id,
-					} )
-				);
-				setOptions( options );
-			} )
-			.catch( ( error ) => {
-				// eslint-disable-next-line no-console
-				console.error( 'Error fetching Gravity Forms:', error );
-			} );
-	}, [] );
 	return (
 		<>
 			<InspectorControls>
@@ -47,13 +40,70 @@ export default function Edit( { attributes, setAttributes } ) {
 												option.value ===
 												parseInt( value, 10 )
 										)?.label || '',
+									prefilledValues: {},
 								} )
 							}
 						/>
 					</PanelRow>
 				</PanelBody>
+				{ ! isLoading && preFillableFields.length > 0 && (
+					<PanelBody
+						title="Gravity Form Pre-Filled Values"
+						initialOpen={ true }
+					>
+						{ errors.length > 0 && (
+							<Notice status="error" isDismissible={ false }>
+								{ errors[ 0 ] }
+							</Notice>
+						) }
+						<Tip>
+							Fields that are marked as “pre-fillable” can be
+							automatically populated with values.
+						</Tip>
+						{ preFillableFields.map( ( field ) => (
+							<div
+								style={ {
+									flex: '1 1 100%',
+									marginBlock: '1rem',
+								} }
+								key={ field.id }
+							>
+								<TextControl
+									__next40pxDefaultSize
+									__nextHasNoMarginBottom
+									label={ `${ field.label } ${
+										field.isRequired ? '(Required)' : ''
+									}` }
+									type={ field.type }
+									value={
+										attributes.prefilledValues[
+											`preFill_${ field.id }`
+										] || ''
+									}
+									onChange={ ( value ) =>
+										setAttributes( {
+											prefilledValues: {
+												...attributes.prefilledValues,
+												[ `preFill_${ field.id }` ]:
+													value,
+											},
+										} )
+									}
+								/>
+							</div>
+						) ) }
+					</PanelBody>
+				) }
+				<ModalControls { ...props } />
 			</InspectorControls>
-			<div { ...blockProps }>{ formTitle } (Gravity Form)</div>
+			<div { ...blockProps }>
+				{ formTitle } (Gravity Form)
+				{ errors.length > 0 && (
+					<Notice status="error" isDismissible={ false }>
+						{ errors[ 0 ] }
+					</Notice>
+				) }
+			</div>
 		</>
 	);
 }