+++ src/blocks/gravity-form-renderer/index.tsx
@@ -3,6 +3,7 @@ import { commentContent } from '@wordpress/icons';
 
 import metadata from './block.json';
 import Edit from './Edit';
+import './style.scss';
 
 registerBlockType( metadata.name, {
 	icon: commentContent,