+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## v2.3.0 - [June 17, 2026]
+
+-   Added: Admin Screen for necessary keys/settings
+
+## v2.2.0 - [June 16, 2026]
+
+-   Block Updates: [Gravity Form Renderer](src/blocks/gravity-form-renderer/CHANGELOG.md),[Modal](src/blocks/modal/CHANGELOG.md)
+
+## v2.1.1 - [June 16, 2026]
+
+-   Fixes to [Gravity Form Renderer Block (now v1.1.1)](src/blocks/gravity-form-renderer/CHANGELOG.md)
+
+## v2.1.0 - [June 15, 2026]
+
+-   Updated: Gravity Form Renderer block now bumped to v1.1.0 and supports the Services contact form
+
 ## v2.0.1 - [June 15, 2026]
 
 -   Fixed: "Vendor" folder is included in plugin zip script