+++ src/blocks/gravity-form-renderer/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 [View the readme](/src/blocks/gravity-form-renderer/README.md)
 
+## v1.2.1 - [June 16, 2026]
+
+-   Fixed: Properly handling recpatcha v3
+
+## v1.2.0 - [June 16, 2026]
+
+-   Added: Block has extra attributes to control modal behavior
+-   Added: Recaptcha notice is included by default
+-   Updated: Error handling and validation is handled
+-   Updated: Block is only allowed once per page
+-   Updated: Pre-Filled values text is now a `<Tip>`
+
+## v1.1.1 - [June 16, 2026]
+
+-   Added: Form clears/resets after 5 seconds
+-   Fixed: Added `isSubmitting` state to form + styles
+-   Fixed: Properly added block attributes to the container
+
+## v1.1.0
+
+-   Added: Block supports pre-filled values
+-   Added: Block uses confirmation settings set in Gravity Forms
+
 ## v1.0.0
 
 -   Init!